Creating of protoytpe scenes, Scene transition script,
The following update is the connection of the menu button and level select buttons. Connecting scenes together and attempting at creating a black fade in and out transition screen. still need to add the winning script after defeating all of the enemies, and automatically transition to the next level or have a UI pop up to let player move on to the next scene.