Sound Manager, Tower Health Bar, Player Health Bar, Respawn Point
Created a sound manager script to put in all audio and sound effects. Connected the Tower Health Bar and Player Health bar to display on their respective UI. Created Respawn point from player to reactor TODO: Respwaning Reactor to a set destination and pullint the player to said location in the event reactor is destroyed.